[E2E] Harden E2E specs that rely on searching filtered search result set
It was noted in https://gitlab.com/gitlab-org/gitlab/-/issues/426248
that under certain circumstances we can see some tests fail if sidekiq
hasn't run some background processes that are required for a project to
appear in the search result page. Under these conditions we should
instead add a retry to the spec, as it isn't expected that the search
should be instantaneous given the dependance on sidekiq here.

This may not resolve every issue, but should help with addressing some
of the observed flaky failures