Update catalog listing queries to use denorm visibility_level column
In previous MRs we denormalized 'visibility_level' in
the 'catalog_resources' table from 'projects'. We also
implemented a syncing process to keep the data in sync
and conducted a data backfill.  In this MR, we update
the Ci::Catalog::Listing queries to use the new
denormalized visibility_level column.  This involves
updating Ci::Catalog::Resource model with visibility
class methods similar to the ones in the Project model.