Add importing: true to GitHub import objects
There are often validations and callbacks on models that should be
switched off during an import.

To disable callbacks on an object we create during an import, we must
always remember to set importing: true on the object before saving, if
that object's class includes the `Importable` module.

This change sets `importing: true` on some objects in the GitHub
importer.

https://gitlab.com/gitlab-org/gitlab/-/issues/432962

Related: https://gitlab.com/gitlab-org/gitlab/-/issues/434114

Changelog: fixed