Add docs about how to finalize BBM
This adds a specific section that describes what is involved in
finalizing a batched background migration. This information is somewhat
scattered around examples and other places right now and it didn't seem
like we had any specific place to put it.

The main reason I wanted to add this doc to begin with was to highlight
that the `gitlab_schema` must not change from queue to ensure. But I
couldn't find any place to put it so I made a new section and then I
also dumped all the other rules I know of in this section.

The `gitlab_schema` rule documented here will become very important in
the next few months as we'll be changing hundreds of tables from
`gitlab_main` to `gitlab_main_cell` so we want to be extra clear about
this.