Enable ClickHouse migrations by default
It turns out that we run migration differently on .com.
See more in https://gitlab.com/gitlab-org/gitlab/-/issues/428126

So this feature flag can't really be tested on .com

On the self-managed we migrate CH only if it's configured:

https://gitlab.com/gitlab-org/gitlab/blob/716a459d47bab3945e6a86c9236680f424fefa68/lib/tasks/gitlab/click_house/migration.rake#L11

Otherwise, we just print a message.

But let's keep the feature flag for a few milestones,
just in case our experimental-customers would want to disable it.

(I'm probably being too paranoid here)