Fix ID dependent jest spec
This jest spec had a hard coded ID, but also relies on fixtures.
Coincidentally one of the IDs in the fixtures ended up being the ID
which was not supposed to exist in the fixture data.

We fix it by generating a unique ID which definitely does not exist in
the fixture (sum of all IDs).