Rework Vite Hot Module Reloading
By setting a different client port we can make the nginx reverse
proxy work properly. The corresponding GDK was merged here:
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3501

We also now start vite on the correct host and we load the
settings directly from `config/vite.gdk.json`.




This is easier than loading it via the environment. We are not requiring
the JSON with an `import` because loading it via readFile is a little
more future proof for ESM.