Hide "Merge immediately" button when fast-forward is not possible
We've recently implemented automatic rebase on merge for fast-forward
merge trains. This had the unintended consequence making the `Merge
immediately` button available when it shouldn't be.

In this MR, we add logic to hide this button when fast-forward merge
trains are enabled and the MR cannot be merged immediately.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/434070

Changelog: fixed
EE: true