Set failure reason for non-persisted pipelines
Makes it so that `pipeline.failure_reason` is accurate for non-persisted
failed pipelines, such as pipelines that have been filtered by rules or
workflow rules. We will later use this to keep track of the source
branch commit sha of the latest valid processed pipeline configuration
for merge requests.

These changes were split out of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/136089 because
that MR is still very large and under development, but in the meantime
we could already merge and fully roll out this code.