Various improvements of initial gitlab-housekeeper feedback
This was followup feedback extracted in
https://gitlab.com/gitlab-org/gitlab/-/issues/435618 for
`gitlab-housekeeper`.

The original implementation of `gitlab-housekeeper` can be found in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/139492 and it is a
local developer tool for automatically generating merge requests for
repetitive tasks like finalizing batched background migrations. It is a
gem you can run from the command line and is being actively developed to
eventually run as a periodic job to save us from many manual repetitive
MRs.