Ensure attempt to complete a verification method is allowed
Users are required to complete verification methods in order depending
on their risk score:

Low: 1. email
Medium: 1. email 2. phone
High: 1. email 2. phone 3. credit card

Previously, phone number verification endpoints were secured by ensuring
that a user is required to complete phone number verification and has
already completed email verification.

Here, the before action hook is made generic so it can ensure that a
verification method is required and all its prerequisite verification
methods are completed before the corresponding action is executed.

For example, if a user is required to complete the following:

Note: that the order of credit card and phone are swapped. This
should not matter.

1. email
2. credit card
3. phone

Hitting the credit card verification endpoints will require email is
method is completed and that credit card verification is actually
required for the user. The same happens for phone number verification
endpoints except for a slight difference: instead of just email method,
completion of both email and credit card methods are checked.