Gem gitlab-housekeeper skip updates if edited by another person
The gitlab-housekeeper gem is a tool for automatically creating merge
requests for repetitive things like finalizing batched background
migrations. Sometimes the generated merge request needs to be amended by
a developer (eg. pipeline fails due to MR title too long). In those cases
we don't want to overwrite their changes.

This merge request extends the gem to look through the system notes on
the MR and detect if the title, description or code has been edited by
anyone else. If that's the case it does not update that particular
detail of the merge request.