Escape unicode null characters before parsing the security report
As the Unicode null character(\u0000) can not be stored by PostgreSQL,
we are escaping it while parsing the security report artifacts by this
change. We will be notifying the user with a warning message shown on
the "pipeline security tab" if the artifact produced by the analyzer
contains a Unicode null character.

Changelog: fixed
EE: true