Skip ingesting the SBOM reports if any of the SBOM builds fail
If a single SBOM job fails, during a pipeline for the default branch,
then the corresponding components are marked as no longer detected.

This is misleading and can lead to false positives and issues further
down the reporting workflow. This [discussion thread][0] discusses the
issues in more detail and comes to the conclusion that we should skip
ingesting the SBOM reports if *any* of these SBOM jobs fails.

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/421666

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/421666#note_1532251397

Changelog: changed
EE: true