[E2E] Harden pipeline with protected variable spec
The specs in pipeline_with_protected_variable_spec rely on adding
a user to a project, but as this operation occurs async I believe
in some situations this may not process in time for the test. By
adding some retry logic here to allow for this we should help to
reduce the number of flaky failures related to this.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/435608
Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/429876