Send GitLab version and hostname during instance activation
Prior to this change, only the activation code, instance identifier and
if the activation is automated was sent during the instance activation.
For upcoming new feature, the GitLab version will be needed to be synced
during the activation.

This change sends info about the instance's GitLab version and its
hostname during the instance's activation to allow the data to be
present from the get go and allow future features.

Changelog: changed
EE: true