[E2E] Harden Group Audit Logs spec
There appears to be a flake in this spec that occasionally occurs
where the test is left requiring us to bypass the 2FA setup flow.
When we enable 2FA, and navigate to a group we are renavigated to
this page where we are promoted to either finish setting up 2FA
or to 'configure it later' - under some circumstances clicking
'configure it later' doesn't get registered, so the rest of the
test flow is impacted and we get stuck on this page. Instead
let's ensure that we use a retry here so so that if navigation
to the group fails, we attempt to 'configure it later' again
which should help avoid flaky failures of this spec setup

Relates https://gitlab.com/gitlab-org/gitlab/-/issues/396793