Rephrase DoD requirement on MR acceptance checklist
We see that the checklist is not always ticked. Those acceptance
guidelines are very important for newer contributors, but are
reviewed by experienced ones only from time to time.

Ticking the checklist does not necessarily ensure that all
the guidelines are followed. Rewording this point let us
make sure the content of the checklist is more important
that the formality of checking it.