Preload saml_provider unconditionally to fix N+1 queries
Preloading saml_providers is required for EE when resolving groups as
saml_provider is used in GroupPolicy to prevent/enable reading the
group.

ForkTargetsResolver now directly uses LooksAhead instead of using it
through ResolvesGroups as we don't want to preload saml_providers for
ForkTargetsResolver because it returns user namespaces in addition to
groups. This results in spec failures because saml_provider is not
present for user namespaces but only for groups.

Changelog: other
EE: true