Refactor repository storage move worker scheduling
The `schedule_repository_storage_update_worker` method is overridden in the `RepositoryStorageMove` model and its subclasses to schedule the appropriate worker. 

The worker is now called with the `id` of the `RepositoryStorageMove` record, instead of passing additional arguments such as the project or snippet ID and the destination storage name. 

This simplifies the code and makes it more maintainable.

Changelog: other
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/139724#note_1700869142