Simplified badges FF
This MR introduces the FF :simplified_badges

For testing purposes we moved the changes from GitLab UI
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3307
directly to GitLab as this is a change which spans over the
hole product and can make a lot of impacts.

Once verified we'll move things to GitLab UI and remove those
overrides again from the product.