Update Slack Notifications removal to 18.0
The Slack Notifications integration is deprecated. We have been
announcing that the integration will be removed in 17.0, however there
are a few blocking issues to its removal (mentioned in
https://gitlab.com/groups/gitlab-org/-/epics/8673) that have not been
done (as of time of writing, in 16.8).

We will extend the milestone we place to remove the Slack Notifications
integration from 17.0 to 18.0 in order to give us time to complete those
features.

https://gitlab.com/gitlab-org/gitlab/-/issues/435906