Change telesign_intelligence feature flag type to ops
telesign_intelligence feature flag controls calls to Telesign's
Intelligence API to fetch risk score and other information for a given
phone number. The feature is stable and will stay on the product as part
of Identity Verification feature but we want to be able to toggle it.

Changelog: changed
EE: true