Populate ci analytics database if ClickHouse is configured
We want to make it easier to set up clickhouse analytics database.
To do that we need to remove some feature flags from
https://gitlab.com/gitlab-org/gitlab/blob/5f571d8c9c7682c126f91a68010728203f25486c/doc/ci/runners/runner_fleet_dashboard.md#L47

This MR removes the feature flag check from in the ingestion service.
Note that FF is still used to guard creation of sync events.
I plan to enable it for self-managed soon, but not remove for some time.

You can see all the proposed changes in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/140333