Remove LfsObjectsWorker overrides of StageMethods
This change refactors ImportLfsObjectsWorker not to override
StageMethods methods, as it was missing out on some of the logic in
that module.

It also contains a few other changes:

- StageMethods no longer finds the project scoped by its import_state,
  in favour of letting some existing logic
  https://gitlab.com/gitlab-org/gitlab/-/blob/d6dea9aa2c2146f12da00fd1734b4b5af0587590/app/workers/concerns/gitlab/github_import/stage_methods.rb#L55-63
  to log and return if the project's import state is #completed? execute
  instead.
- It no longer logs "stage finished" if the worker hits a rate limit and
  reschedules itself to retry later. Instead it logs "stage retrying
  after rate limit".
- It moves queuing the RefreshImportJidWorker out of try_import, to keep
  that method purely about handling rate limit exceptions in #import.

To catch if any behaviours of StageMethods are overridden,
the tests of the module have been ported to the existing shared examples
group, so we run them per stage importer.