[E2E] Harden cancel pipeline when user blocked spec
In some situations the async nature of how users are added to a
project can mean that the permissions haven't been applied prior
to a request being made in our test. We should add a retry to
allow for this situation to avoid flaky test failures in these
situations.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/417661