Exclude artifact & package sizes when calculating project export limit
Whenever Project Export is initiated, we check if the export size 
exceeds the limit. For some projects that have a lot of 
packages/artifacts, the export would fail. But Project Export doesn't
include packages/artifacts in its contents, so there is no need to 
include those sizes in the calculation.

Changelog: fixed