[E2E] Harden group access token spec
This appears to be caused by cases where sidekiq doesn't run jobs
in the background fast enough for the newly created group access
token to be asoociated to the projects, leading to a 403 status.

By adding an additional background check here, we can verify that
the token has been successfully associated to the project prior
to executing the tests in question.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/433393
Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/431378