Queue workers only for objects with attachments
During a GitHub import, when we queue a worker we spread them to run
1,000 per minute. The more workers queued, the longer the spread across
time.

Because most workers will do nothing, this spread adds unnecessary time
to an import. The larger the import, the more time is wasted in this
stage.

This change will queue workers only for objects that have attachments.

https://gitlab.com/gitlab-org/gitlab/-/issues/436400