Save external_identifiers on ImportFailure for file import relations
The `FileImport` model has a field for tracking the external identifier
associated with a failed record import. This field is used for other
importers such as GitHub and BitBucket, but not file-based import.

This MR tracks the IID associated with issues, merge requests,
milestones and CI pipelines that fail to import. It also lays the
groundwork for tracking identifiers for other relations, should the
need arise and we can identify a unique key.

Changelog: changed