Emit event when a release is published
Previously, there was no event to which to subscribe to if we wanted to
add logic both 1/ when a release is created with not setting a different
release date and 2/ when the release date of a pre-existing release
record has come.

This commit introduces the `::Projects::ReleasePublishedEvent` event in
`EventStore`, which can be subscribed to implement logic when a release
happens.

Changelog: added