Group Dependency list: use backend error message
When we encounter an error while fetching the group dependency list,
in many cases the backend returns a detailed error message.

Currently we are only showing a generic, hardocded error message
in the UI. This commit changes it, so we we leverage the
message from the backend if it is present.