Refactor ApplicationSetting specs to use validate_numericality_of
Instead of using allow_value or not_to allow_value using
validate_numericality_of results in concise specs.

The `validate_numericality_of` matcher checks if the argument is a
string or a number so we don't need to test it explicitly in the tests.
The common examples are now under the same context using table syntax.