Use GDK's .ruby-version as canonical settings in build-gdk-image
This follows the change made in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/132816.

The `build-gdk-image` job attempts to install GitLab with a given GDK
version. If the Ruby version is updated in the `gitlab/` directory,
the GDK pre-flight checker will detect that the GDK Ruby version is
inconsistent.

To avoid build failures that result from this, use the GDK
`.ruby-versions` as the canonical file by copying it into the `gitlab`
directory since that's what's actually installed.

This will help avoid causing broken builds whenever `.tool-versions`
change.