Add auto_cancel:on_new_commit to canceling redundant pipelines
This is the fourth step of full implementation of
workflow:auto_cancel:on_new_commit and using it in the canceling
redundant pipelines feature.

In this change, we are using auto_cancel_on_new_commit in
CancelRedundantPipelinesService and adding safe_cancellation to
CancelPipelineService.

These changes are behind the FF ci_workflow_auto_cancel_on_new_commit.