Remove database event tracking
We introduced tracking of database events via Snowplow through multiple
iterations, most recently in
https://gitlab.com/gitlab-org/gitlab/-/issues/390827. This was done to
test out whether Snowplow could be used to create a replica of our
database.

Within https://gitlab.com/groups/gitlab-data/-/epics/880#note_1474689144
it's described that the PoC led to discarding Snowplow as a viable
option for data replication. This is also confirmed in
https://gitlab.com/gitlab-org/gitlab/-/issues/390827#note_1562890019.

Now this is just dead code which can lead to confusion.