Update uri gem to v0.13.0
If a user installs the latest version of uri (v0.13.0 as of this
writing), then Ruby will fail to run:

```
`check_for_activated_spec!': You have already activated uri
0.13.0, but your Gemfile requires uri 0.12.2. Prepending `bundle exec`
to your command may solve this. (Gem::LoadError)
```

As explained in https://github.com/rubygems/rubygems/issues/7178,
this only appears to happen when:

1. A local gemspec is used.
2. The gemspec has a home page URL.

RubyGems loads the latest version of `uri`, but this conflicts
with a slightly version required in this `Gemfile.lock`.

The only workaround I can see is to update the uri gem. Fortunately,
the changes are mostly cleanup:
https://my.diffend.io/gems/uri/0.12.2/0.13.0

Changelog: changed