Adds tooltip to clarify container image storage
In some situations, we may present a non zero repository size
when some atomic manifests are not tagged
(so they don't show up in the list of images)
but are referenced by a tagged manifest list/index (which size is zero).
Thus, this leads to a situation where the sum of the size
of listed images does not match the displayed size of the repository.

Changelog: changed