Add allowlist to db/docs files for cross-joins/fks/txns
As part of the
[Cells](https://docs.gitlab.com/ee/architecture/blueprints/cells/)
effort we need to classify all tables as `gitlab_main_cell` or
`gitlab_main_clusterwide`. In the future these tables will be split into
different databases. For now when we classify these tables we will start
to get errors in CI for
[cross-joins](https://docs.gitlab.com/ee/development/database/multiple_databases.html#removing-joins-between-ci-and-non-ci-tables),
[cross-transactions](https://docs.gitlab.com/ee/development/database/multiple_databases.html#removing-cross-database-transactions)
and
[cross-database foreign keys](https://docs.gitlab.com/ee/development/database/multiple_databases.html#foreign-keys-that-cross-databases).

This MR adds an allowlist in the `db/docs` file per type to avoid seeing
these failures once we classify the table as `gitlab_main_cell`. This is
to to support
[automatically labelling hundreds of tables](https://gitlab.com/gitlab-org/gitlab/-/issues/429172)
in bulk with `gitlab_main_cell` (as part of the
[Organization
Isolation](https://docs.gitlab.com/ee/architecture/blueprints/organization/isolation.html)
effort)
. This will result in lots of pipeline failures
and we want to deal with these those cross-join (etc.) fixes later. For now
we just want to classify things.

You can see the example format of allowing in the
`db/docs/achievements.yml` change which was chosen as an arbitrary
example to prove that this works.