Clarify what #cannot_be_merged? == true means
cannot_be_merged? is generally indicative of conflicts, and is set via
MergeRequests::MergeabilityCheckService. However, it can also indicate
that either #has_no_commits? or #branch_missing? are true.

Here, we update an error message that has caused some customer confusion
(thinking there was a conflict when the issue was an empty MR..), update
API docs, and sprinkle in some code comments for additional context or a
future cleanup effort.