Use JSONSchema validation errors in EventStore data error
Prior to this change, `Gitlab::EventStore::Event` would validate event
schema but not reveal any information about the validation errors
themselves.  The underlying library actually reports all of this
information if we use a the `#validate` method, and we can then use the
`JSONSchema::Errors.pretty` to get human readable output for the error
message.

In this change, we replace the actual schema with just the human
readable errors. This makes things easier to debug and means that each
distinct type of invalid event gets its own error message. This should
be safe to output all the time. It only reports the error's type and
location in the event; it does not include any event data.