Fix JSON schema of Ci::JobArtifactsDeletedEvent
I had originally picked this event out to provide an example for
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/141110, but found
that it always passed my validations, no matter what type of an element
I put in there. Turns out the array item type declaration was invalid,
but the schema validation passed because the library ignores the unused
`'properties'` key for `'type' => 'array'`.

This change should be safe because the data is
`artifacts.map(&:job_id)`, and the `job_id` column a non-nullable
integer.