Fix blank values handling in the dependency proxy
In the dependency proxy for packages, the settings object can manage a
username and a password. Those values are expected to be string values
but due to how model validations and database validations worked, it was
possible to try to insert invalid values.

Blank values for these columns are now consistently nullified.