Replace signin-container for a specific selector
A div element with the `signin-container` was use for a JavaScript
functionality to preserve the URL fragments and the `remember_me` query.

In this MR we replace the `#signin-container` by a more specific
`.js-non-oauth-login` selector.

We also move the `.js-oauth-login` selector to the parent element so
both, `js-oauth-login` and `js-non-oauth-login`, works similarly.