Swap babel proposal for transform plugins
Once a TC39 proposal becomes part of the ECMAScript spec, babel
deprecates the corresponding `@babel/plugin-proposal-*` in favour of a
`@babel/plugin-transform-*`.

We have been enabling a hardcoded list of plugins. Some for
compatibility with webpack@4, some for unknown reasons. We should double
check whether we need them, but that should be done in a separate MR.

That being said, all plugins we use made it to the ECMAScript standard
and we can happily use them as transform plugins.