Fix group milestone deletion
Group milestones fail to delete with an internal server error
if a merge request is assigned to it.

The `MergeRequests::BaseService` expects a projects as parameter.
The `Milestones::DeleteService` however passed a group as the project parameter.

That caused a `NoMethodError` since the `@project` variable is not
being set in the `BaseContainerService` because the passed container
is a group. That resulted in the `project` attr_accessor returning nil.

Changelog: fixed