Ignore pending access requests when calculating the highest group role
Currently, inviting a group member to a subgroup fails with reporter or
guest access if there's a pending access request of the same user in the
parent group. This happens because when we calculate the highest group
member access we don't ignore the pending access request. This change
fixes this issue by just ignoring the pending access request as users
with pending access requests are not members of the group.

This issue also affected the `/api/v4/projects/:id` API which has also
been fixed.

Changelog: fixed