Fix "error on credentials page for PATs with deleted group/project"
As project is destroyed related user deletion takes place as the
same happens async. Token deletion happens as the part of User
deletion. That means even when project is deleted, token is there
in database for some minutes around that time if user tries
to access admin/credentials page an error is thrown. With this MR
we will display the status as Deleted to admin in case the token
does not belong to Group or Project

Changelog: fixed
EE: true