Do not create AI summary for MR with empty extracted diffs
When we send a request to AI to summarize merge request but the
diff is empty, we send the prompt with an empty diff to summarize
and it can respond with irrelevant summary.

To prevent that, we don't make a request to AI when diff is empty
and if the extracted diff we use for the prompt is blank (e.g. diff
contains binary files only).