REFACTOR ActivityPub serializers
ActivityPub is all about JSON : it's used in endpoints to display
information about various actors (the ActivityPub term for any kind of
resource that generates a feed), as well as in messages exchanged
between servers about what those actors are doing ("activities", in
ActivityPub terminology). For that reason, one of the very first thing
we did when implementing ActivityPub was creating serializers for it, in
the `ActivityPub` namespace, in the merge request
[about displaying the first actor's profile and outbox](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/127023).
Those serializers are only used by ActivityPub features behind a
feature flag, for now.

While working on additional features, though, it was made clear that the
architecture for those serializers was not going to work for everything.
First, an actor and an activity usually have different boilerplate, so
it helped distinguish them through an ActorSerializer and an
ActivitySerializer, both inheriting from an ObjectSerializer (referring
to the ActivityStreams' [Object type](https://www.w3.org/TR/activitystreams-vocabulary/#object-types)).

Second, using `BaseSerializer` plural features does not work here.
ActivityStreams paginated collections are
[their own beast](https://www.w3.org/TR/activitystreams-core/#paging).
They are linked lists, with a dedicated document serving as index and
providing links to the various resources - it just does not fit what
BaseSerializer allows to do. So an other serializer,
`ActivityPub::CollectionSerializer` has been added to handler them. This
simplifies a lot previous the architecture that was fighting to make
this fit `BaseSerializer` plural feature through a lot of overriding.

This commit refactors the ActivityPub serializers to reflect that and
pave the way for the features coming after. It does not affect existing
code (the API used so far for the releases actor is unchanged).