Use dependent: :destroy for projects deployments
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105333 introduces
the solution and path forward for the problem of query cancellation due
to timeouts related to project deletion. We see this issue happening
with deployments in https://gitlab.com/gitlab-org/gitlab/-/issues/409828

This commit follows the same pattern, leveraging the existing work and
the usage of destroy_dependent_associations_in_batches in
Projects::DestroyService.