Align icon with text in Code Quality widget
The container has `align-items: baseline` set, but there was an empty
element being rendered between the icon and the text of the Code Quality
finding, which interfered with determining _where_ the baseline was.
That element contains the header, if given, and a help popover, if
given. When neither are given, the element shouldn't be rendered.

Note that the rendering logic of this component is fairly convoluted.
The `StatusIcon` is rendered in two different places, depending on
whether or not the `header` prop is truthy (why not consider the slot as
well?). This should probably be simplified in future to avoid
regressions like this.

Changelog: fixed