Store full include path for CI components
The `path` field on `ci_components` currently holds the file path of the
component within the component project repository. However, we aren't
currently using that data. Instead, we need to provide the include path
for the component so we can display it on the catalog resource details
  page. This commit stores the include path instead of the file path.

Existing component records will still have the file path. We'll decide
later whether it is necessary to update them with a data migration.

Changelog: changed