Stream summarized notes
This enables streaming of summarized notes and provides the capabilities
to render the streamed chunks on the frontend.

Streaming only works with Anthropic and the
`summarize_notes_with_anthropic` feature flag being turned on.

The FE change is backwards compatible during deployment as we already
pass through the `clientSubscriptionId` from the mutation to the
broadcasting of the subscription.